Prepare for games management
