+++ Assets/Scripts/Cgs/Menu/GamesManagementMenu.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 07e97a02d62f0e94f8638201d8e1343b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
