+++ Assets/Scripts/Cgs/Menu/MainMenu.cs
@@ -2,32 +2,20 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-using System;
 using System.Collections.Generic;
 using JetBrains.Annotations;
 using UnityEditor;
 using UnityEngine;
-using UnityEngine.Events;
 using UnityEngine.EventSystems;
 using UnityEngine.SceneManagement;
 using UnityEngine.UI;
-using UnityExtensionMethods;
 #if !UNITY_ANDROID && !UNITY_IOS
-using SimpleFileBrowser;
 #endif
 
 namespace Cgs.Menu
 {
     public class MainMenu : MonoBehaviour
     {
-        public const string ImportGamePrompt = "Download from Web URL,\n or Load from ZIP File?";
-        public const string DownloadFromWeb = "Download from Web URL";
-        public const string LoadFromFile = "Load from ZIP File";
-        public const string SelectZipFilePrompt = "Select ZIP File";
-
-        public const string DownloadLabel = "Download Game";
-        public const string DownloadPrompt = "Enter CGS AutoUpdate URL...";
-
         public static string WelcomeMessage => "Welcome to CGS!\n" + WelcomeMessageExt;
 
 #if UNITY_ANDROID || UNITY_IOS
@@ -58,10 +46,6 @@ namespace Cgs.Menu
 
         private const float StartBufferTime = 0.1f;
 
-        public GameObject cardGameEditorMenuPrefab;
-        public GameObject gameImportModalPrefab;
-        public GameObject downloadMenuPrefab;
-        public GameObject gamesManagement;
         public GameObject versionInfo;
         public Text currentGameNameText;
         public Image currentCardImage;
@@ -74,25 +58,6 @@ namespace Cgs.Menu
         public Button joinButton;
         public GameObject quitButton;
         public Text versionText;
-
-        private CardGameEditorMenu CardGameEditor =>
-            _cardGameEditor ??= Instantiate(cardGameEditorMenuPrefab).GetOrAddComponent<CardGameEditorMenu>();
-
-        private CardGameEditorMenu _cardGameEditor;
-
-        private DecisionModal ImportModal =>
-            _importModal ??= Instantiate(gameImportModalPrefab).GetOrAddComponent<DecisionModal>();
-
-        private DecisionModal _importModal;
-
-        private DownloadMenu Downloader => _downloader ??= Instantiate(downloadMenuPrefab)
-            .GetOrAddComponent<DownloadMenu>();
-
-        private DownloadMenu _downloader;
-
-#if UNITY_ANDROID || UNITY_IOS
-        private static string _zipFileType;
-#endif
 
         private void OnEnable()
         {
@@ -109,13 +74,8 @@ namespace Cgs.Menu
 #else
             quitButton.SetActive(false);
 #endif
-            gamesManagement.SetActive(false);
             versionText.text = VersionMessage;
             versionInfo.SetActive(false);
-
-#if UNITY_ANDROID || UNITY_IOS
-            _zipFileType = NativeFilePicker.ConvertExtensionToFileType("zip");
-#endif
 
             if (!HasSeenWelcome)
                 CardGameManager.Instance.Messenger.Ask(WelcomeMessage, DeclineWelcomeMessage, AcceptWelcomeMessage,
@@ -187,7 +147,7 @@ namespace Cgs.Menu
             else if (Inputs.IsSave)
                 EditDeck();
             else if (Inputs.IsFocusBack && !Inputs.WasFocusBack)
-                ToggleGameManagement();
+                ShowGamesManagementMenu();
             else if (Inputs.IsFocusNext && !Inputs.WasFocusNext)
                 ExploreCards();
             else if (Inputs.IsOption)
@@ -211,23 +171,10 @@ namespace Cgs.Menu
         }
 
         [UsedImplicitly]
-        public void ToggleGameManagement()
-        {
-#if !UNITY_WEBGL
-            if (Time.timeSinceLevelLoad < StartBufferTime)
-                return;
-            gamesManagement.SetActive(!gamesManagement.activeSelf);
-            versionInfo.SetActive(gamesManagement.activeSelf);
-            EventSystem.current.SetSelectedGameObject(null);
-#endif
-        }
-
-        [UsedImplicitly]
         public void SelectPrevious()
         {
             if (Time.timeSinceLevelLoad < StartBufferTime)
                 return;
-            gamesManagement.SetActive(false);
             CardGameManager.Instance.Select(CardGameManager.Instance.Previous.Id);
         }
 
@@ -236,81 +183,17 @@ namespace Cgs.Menu
         {
             if (Time.timeSinceLevelLoad < StartBufferTime)
                 return;
-            gamesManagement.SetActive(false);
             CardGameManager.Instance.Select(CardGameManager.Instance.Next.Id);
         }
 
         [UsedImplicitly]
-        public void CreateNew()
+        public void ShowGamesManagementMenu()
         {
+#if !UNITY_WEBGL
             if (Time.timeSinceLevelLoad < StartBufferTime)
                 return;
-            CardGameEditor.Show();
-        }
-
-        [UsedImplicitly]
-        public void Import()
-        {
-            if (Time.timeSinceLevelLoad < StartBufferTime)
-                return;
-
-            ImportModal.Show(ImportGamePrompt, new Tuple<string, UnityAction>(DownloadFromWeb, ShowDownloader),
-                new Tuple<string, UnityAction>(LoadFromFile, ShowFileLoader));
-        }
-
-        private void ShowDownloader()
-        {
-            Downloader.Show(DownloadLabel, DownloadPrompt, CardGameManager.Instance.GetCardGame, true);
-        }
-
-        private static void ShowFileLoader()
-        {
-#if UNITY_ANDROID || UNITY_IOS
-            var permission = NativeFilePicker.PickFile(path =>
-            {
-                if (path == null)
-                    Debug.Log("Operation cancelled");
-                else
-                    CardGameManager.Instance.ImportCardGame(path);
-            }, _zipFileType);
-            Debug.Log( "Permission result: " + permission );
-#else
-            FileBrowser.ShowLoadDialog((paths) => CardGameManager.Instance.ImportCardGame(paths[0]),
-                () => { }, FileBrowser.PickMode.Files, false, null, null,
-                SelectZipFilePrompt);
+            Debug.Log("wire it up");
 #endif
-        }
-
-        [UsedImplicitly]
-        public void Sync()
-        {
-            if (Time.timeSinceLevelLoad < StartBufferTime)
-                return;
-            CardGameManager.Instance.StartCoroutine(CardGameManager.Instance.UpdateCardGame(CardGameManager.Current));
-        }
-
-        [UsedImplicitly]
-        public void Edit()
-        {
-            if (Time.timeSinceLevelLoad < StartBufferTime)
-                return;
-            CardGameManager.Instance.Messenger.Show("Edit is Coming Soon!");
-        }
-
-        [UsedImplicitly]
-        public void Delete()
-        {
-            if (Time.timeSinceLevelLoad < StartBufferTime)
-                return;
-            CardGameManager.Instance.PromptDelete();
-        }
-
-        [UsedImplicitly]
-        public void Share()
-        {
-            if (Time.timeSinceLevelLoad < StartBufferTime)
-                return;
-            CardGameManager.Instance.Share();
         }
 
         [UsedImplicitly]
